Added Contour-Class
The Contour-Class represents a contour of a cluster.
The contour is represented as a set of points, which will form a curve when linked together that describes the contour of the cluster.
It was implemented for increasing the contour (e.g. by 10%) and then check whether some clusters intersect with each other