Bug fixed: Clusters with 0 points
A point was not added to the cluster after creating a new cluster for the point.
The reason was a bug that ended up in an infinite loop when trying to process the same point again.
An easy way would have been to add the point manually, but this would not have fixed the bug.
The bug lied in normal.dot(pointNormal) where the dot-product was not within [-1,1] despite both vectors beeing normalized.
The reason for this was floating point precision, i.e. the result of the dot-product of one datapoint was very slightly above 1, and then the acos resulted in a -nan(ind).