Added intersections of clusters
Each combination of 3 clusters is intersected.
If the intersection point is within a certain range (max_distance + 10) for each cluster, the point is considered as valid and added to "intersection-points".
The intersectionpoints are the "edges" of the room.

Note: "&loaded_points" was removed from "Cluster" because erasing of Clusters from a vector was not possible because of it.