Finished Interpolation, Added Planification
Interpolation:
Each combination of 3 clusters are intersected which leads to intersection points.
These can be used later on to retrieve exact cornerpoints.

Planification:
Each cluster is filled up with quadratic planes.
For this, the clusterpoints are first transformed to 2 dimensions.
There, an algorithm is applied to fill up the space where the point density is high enough.
In the end, the result is transformed back to 3 dimensions.