Improved 2D to 3D transformation
For the 2D to 3D transformation, the defined axis were required to transform the plane back to 3 dimensions. Currently, this was done by just recalculating the axis the same way as it was done for 3D to 2D.
This is very errorprone and also slower in computation because the 2D to 3D transformation was in a loop and the axis had to be computated many times for just one cluster.

Now, the axis are computated separately and passed as parameter to both 3D to 2D and 2D to 3D.