Added edge intersection, wall merging
Edge intersection:
The walledges are now intersected s.t. we get a closed room.
Doesn't work perfectly yet.

Wall merging:
Some wall clusters could be merged into one cluster. This is because at generating, the normal and center are not "perfect" yet. This may lead to multiple clusters when there should only be one.
Wall merging is currently disabled because it may contain bugs.
Furthermore, the results were not as good as expected.