Improved Clustering, added cubes, added triangles
Clustering was improved. Merging works now and absorbing was also added.
Merging merges two clusters with the same normal.
Absorbing is even stronger, where the normal is ignored if most points
of a smaller cluster can also seen as members of a bigger one.

Cubes were added, they are not used (yet) however.
They are part of a second approach which intended to first build
a "big" volumetric shape, which is then cut by the planes.
The idea would be like cutting a stone to get the exact room.
However, it was not developed further.
Missing is the rather difficult part of actually cutting the cubes and
merge cubes / extract bigger contour-polygons.

The triangle-Approach intersects every 3 clusters that are not rather
perpendicular (an angle <12.5 is seen as "rather perpendicular")
to each other.
An intersection yields an intersection-point and 3 lines.
The 3 lines are then used to get 3 arrows which describe the contour.
For example, a corner in the room would yield 3 arrows directing inwards.


For now, the triangles are used to get closed circles within the plane.
The initial idea was to get a closed circle of one cluster and then
continue on the adjacent clusters to the closed circle.
However, a cluster might yield several closed circles (for example, a
roof that is divided by a horizontal bar).
There is also a possibility that the circle is not correct and we
get stuck later (for example, going through a furniture).
We would need to go back and adapt the circle.
The current idea to this challenge is not implemented yet.
It operates on all possible triangles at once, not only withint a cluster.
The idea is to first compute all triangles, and then use a tree where each
node has 3 subnodes representing the possible ways to go for each arrow.