Added recursive 3-dimensional triangle build
Replaced IntersectionTriangle with the new TriangleNode3D class.
They contain the same information, but are also able to represent a tree.

Before the recursion, all possible connections to other triangles are stored in
possibilities (possibilities[0] for arrow0 connections, and so on).
The recursion iteratively tries to connect open arrows (chosen = -1) to other triangles.
When a collision happens or a triangle can not be connected to anything, it rolls back the connections and tries another way.
When all ways are tested and we didn't close all arrows, it returns false.
On the other hand, if all open arrows are closed, it returns true.
The Nodes got changes and all have unique connections stored in "chosen".

Added also code to extract clusters with the respective polygon.