Added Mesh generation, improved triangulation approach
A mesh is now generated for the first resulting triangulation.

The triangulation was improved a little bit.
For example, for two triangles to connect, two arrows need not only to have the
opposite direction, but they must also "intersect in the middle".
The possibilities are now sorted, s.t. the closest intersection triangle is chosen first.
The calcArrowScore was also improved, s.t. clusters with multiple point-parts are better divided. It still doesn't work very well. The main issue is that it iterates in a square (it increases dist for all axis equally.). However, some clusters are more rectangular and there arises the issue that the required occupation is not met because we added to much empty squares.