Implemented a new Triangle-finding algorithm
The old triangulation-finding algorithm worked by looking at each cluster separately and suggest a quadrant. If the quadrants fit together, an triangle was added.
It had two drawbacks:
 - Iterating from center in a quadratic way doesn't adopt to rectangular-shapes.
   Because of that, the required occupation needed to be very low, resulting in
   unwanted triangles or not catched triangles.
 - It sometimes added occupation from further away and didn't look at the
   other arrows that there was no supporting wall from the other arrows.
   This too leaded to unwanted triangles or flipped triangles and required the
   minimum occupation to be higher in contrast to the first problem, resulting in a
   very narrow or even overlapping minimum Occupation Requirement.

The new triangulation iterates over each possible arrow-combination and returns the best 3 arrows and lengths. Because of the lengths, we can also store the suggested lengths in the triangle. With this, we can order followers from a triangle not by distance of the centers, but by the "invented"/"empty" length.
The new triangule-finding algorithm has one drawback. It is very slow!
Because of this, the searching lenght is reduced to 2 meters, and the patch sizes are now 20cm. It could maybe be improved by not sotring the falt quadrats in boolxbool array, directly add up the occupied qudrats. This would then represent an arrow which is divided into parts, and each part has 2 integers which represent the number of occupied qudrats. It could maybe also further optimized by quitting 3 arrow combinations early based on the occupations and maybe also based on iterations before.