Added fast forwarding for triangles, Removed absorbing
The triangle-calculation was improved by deciding faster whether 3 arrows are
correct by looking at the first 2 arrows. If the third arrow doesn't change the result,
we skip calculating it.

The clustering algorithm was improved by removing the absorbing step and improved the merging-algorithm instead. Absorbing was basically just a merging with a very small angle-fraction (it ignores the normal).
Absorbing also looked at every point and not just the cluster center like in merging. This functionality was added to merging and now absorbing could be removed.