Refactoring: Put the triangle-finding in "TriangleAttempt", other changes
Other changes were necessary to put the triangle-finding in its own class.
The main side-effect is that the Cluster-class is now in its own files.

The fast-forwarding in the arrow-calculation was removed to prepare already
for the new arrow-iteration, which goes from 0,0 and grows by distance.
The difference to earlier algorithms is that for each distance, each possible combination of x,y is tested.
There shouldn't be more tried combinations, but the advantage is that we can quit early if the distance is for example bigger that 1 meter.
This has the advantage of a faster computation and should yield better results (which was the main reason for this change).
It should yield better results because in some cases, bigger arrows were chosen because they had higher occupations, but smaller possible arrows closer to the center were actually wanted. This change could also have been made using a function altering the score to give more value to smaller arrows. But I think quitting early if a result is valid is better since the score-altering-function would likely not work in every case, and an additional parameter of how much more "value" we want to give closer arrows would be necessery.
In any case, this refactoring is beneficial even if we later decide to use the old arrow-iteration again.