Improved arrow-finding-algorithm, other changes
The arrow-finding-algorithm (aka TriangleAttempt), we iterate now from 0,0 and
try every combination of arrows. Because of that, we can now quit early and give more importance to arrows closer to the center. The program does not run faster as expected, because most of the arrows need to be computed fully anyway because they are invalid (so we iterate to the border to potentially find anything).
Note that because of the early quitting, the arrows may not have their perfect score.
The early quitting could be replaced by an "arrow"-lock, s.t. if we find anything, we may not change our arrow-directions later.

The ordering of possible triangle-followers was changed, now it is more likely that we  find the correct "next follower". Because of that, the "searchWidth" could now be increased to 4 meters and we still get the correct room approximation for the "white room".

Note:
A bug was solved in the "distIteration"-function, where the second and third surface-iterator had a wrong value for dir1_min and dir1_max.
There was also a second bug in distIteration, where for each surfaceIterator-callback, computeArrowCombiOcc was only executed for the positive dist value.