Implemented cache for arrow-calculation, other improvements
The arrow-finding-algorithm now has a cache per TriangleAttempt where cluster-occupations can be stored. This increased the speed by around 4-5 times (of course, this is just with the current pointcloud and current settings like "searchWidth=4". I'm not sure how the time complexity is affected, but I think it's more than just linearly faster). Furthermore, for flatQuadrats and max_npA, object attributes are now accessed which made it also faster.

Added also sorting of the clusters according to number of points, to first merge big clusters with smaller clusters.

Merging now takes requireCloseness as double to accept more precise merging-tasks. However, this feature is not fully used yet.