Improved / Changed many things
 - PCA can now be computed separately on a set of points and not just on the neighbors of one point.

 - PCA-max-radius is now set to the parameter "max_dist" directly. This makes it easier to control the extend of the PCA.

 - A new merging approach was tested where each cluster has a maximum positive and negative extend. This is the maximum distance away from the plane a point in the cluster has. If points of another Cluster fit all into the cluster, the other cluster gets merged. It is now commented out because it did not work as expected. It is clearer and more robust if the extend is set manually, otherwise it may get too high.

 - PCA-dist, PCA-K, proportion and required angle are now all taken from the required supported minimum wall-length and required supported wallirregularities.

 - Fixed bug in TriangleAttempt in "allow_early_quit_after".

 - Improved Triangle-linking-algorithm (TriangleNode3D).
First, if a recursion didn't work out, it was not completely rolled back. This is now fixed by introducing "ModificationElement" and "RecursionModifications".
Second, a limit for arrows is now set. This limit is the distance until we reach a triangle that goes into the same direction as us. This triangle could be one generated from furniture (which can be ignored), but it could also be from a wall. So what we do is that we first try all possibilities which don't have to go past the limit. If we don't find anything, we allow our following triangles to go past the limit.
Third, the score was improved s.t. if two triangles are close enough, we look at their score instead of the distance.

 - Cluster generation was made much faster by suggesting clusters to neighbors. If one point creates a cluster or is added to a cluster, it suggests its neighbors to also go to this cluster. The neighbors then first try to go to this cluster and only if they don't fit in, they try all other clusters.