Re-implemented best-fitting-plane for clusters, added video for triangles
A former commit added a PCA-calculation for clusters to compute the best-fitting-plane. It was removed because it didn't seem to work. However, there was likely a bug where the best-fitting-plane was only calculated for copies of the clusters.
It is now re-added and the best-fitting-plane is computed before merging for every cluster and after each merge.
After this change, the results are now more stable especially across different point densities.

Video ply-files can now be created for the triangle-finding-algorithm.
It is commented out to avoid unnecessary computations.