Refactoring: Separated code in "wallextender" to mainly 3 files
The extendWallpoint-function was a function which should initially only find one cluster by extending / growing from a certain point.
It ended up doing all the computations, namely "clustering", "triangle-finding" and "triangle-linking".
The code was separated and put in the respective files:
The "clustering" to the file "Cluster.cpp"
The "triangle-finding" to the file "TriangleAttempt.cpp"
The "triangle-linking" to the file "TriangleNode3D.cpp"

The files "SpecialPrinter" were created which are able to print/write some special datatypes (like clusters or triangles).