Merging a cluster with a required percent now adds all points
Until now, it just added the points which more or less fit in the new cluster.
This was done to avoid changing the cluster-normal to much after a merge.
But since the cluster-normals are now calculated using the best-fitting-plane,
it makes more sense to add all points of the cluster that gets merged.

Other small changes.